Rotación de Refresh tokens

+++ migrations/20260105163850-create-vino.js
@@ -22,7 +22,7 @@ module.exports = {
         allowNull: false
       },
       precio: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL(10, 2),
         allowNull: false
       },
       stock: {
